Lock ID3 to version 2.3 (#35)
* Lock ID3 to version 2.3

* Move ID3v2 settings to SettingsService

* Improve reading settings

* Move settings files
