+++ AudioTagger.Console/Program.cs
@@ -4,6 +4,7 @@ global using System.Collections.Generic;
 global using System.Collections.Immutable;
 global using System.IO;
 global using AudioTagger.Library.MediaFiles;
+global using AudioTagger.Library.Settings;
 using Spectre.Console;
 using System.Text.Json;
 
@@ -43,17 +44,17 @@ public static class Program
             return;
         }
 
-        Settings settings;
-        var settingsResult = SettingsService.Read(printer, createFileIfMissing: false);
-        if (settingsResult.IsSuccess)
+        var readSettingsResult = SettingsService.Read(printer, createFileIfMissing: false);
+        if (readSettingsResult.IsFailed)
         {
-            settings = settingsResult.Value;
-        }
-        else
-        {
-            printer.Error(settingsResult.Errors[0].Message);
+            printer.Error(readSettingsResult.Errors[0].Message);
             return;
         }
+        Settings settings = readSettingsResult.Value;
+
+        SettingsService.SetId3v2Version(
+            version: SettingsService.Id3v2Version.TwoPoint3,
+            forceAsDefault: true);
 
         Queue<string> argQueue = new(args.Select(a => a.Trim()));
 
@@ -103,7 +104,7 @@ public static class Program
 
         if (!filesData.Any())
         {
-            printer.Error("No files found.");
+            printer.Print("No files found.");
             return;
         }
