+++ AudioTagger.Library/Settings/Settings.cs
@@ -1,0 +1,30 @@
+using System.Text.Json.Serialization;
+
+namespace AudioTagger.Library.Settings;
+
+public sealed record Settings
+{
+    [JsonPropertyName("duplicates")]
+    public Duplicates Duplicates { get; set; } = new();
+
+    [JsonPropertyName("tagging")]
+    public Tagging? Tagging { get; set; }
+
+    [JsonPropertyName("renamePatterns")]
+    public ImmutableList<string>? RenamePatterns { get; set; }
+
+    [JsonPropertyName("artistGenres")]
+    public Dictionary<string, string>? ArtistGenres { get; set; } = new();
+}
+
+public sealed record Duplicates
+{
+    [JsonPropertyName("titleReplacements")]
+    public ImmutableList<string>? TitleReplacements { get; set; }
+}
+
+public sealed record Tagging
+{
+    [JsonPropertyName("regexPatterns")]
+    public ImmutableList<string>? RegexPatterns { get; set; }
+}
